.github/fonts/erda-ccby-cjk/requirements.txt
# ERDA CC-BY CJK Font Generator Dependencies
# Python 3.11+

# Font generation
fonttools>=4.47.0

# Configuration
PyYAML>=6.0.1

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code quality
black>=23.0.0
ruff>=0.1.0
